Updated folder structure within Components
