+++ src/components/Inventory_Components/Cards/Cards.js
@@ -5,7 +5,7 @@ import TotalQuantity from './Total_Quantity_Card'
 import CovidProducts from './Covid_Products_Card'
 import LowStock from './Low_Stock_Card'
 
-import '../../App.css'
+import '../../../App.css'
 
 
 // ----------------------------------------------------------------------
