+++ src/components/Inventory_Components/Form/FormDesign.js
@@ -2,7 +2,7 @@ import { Grid, Container } from '@material-ui/core';
 import InsertForm from './Product_Insert_Form';
 import ProductTable from '../Table/ProductTable';
 
-import "../../App.css";
+import "../../../App.css";
 
 
 // ----------------------------------------------------------------------
