+++ src/components/Inventory_Components/InventoryDashboard/InventoryDashboard.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../../App.css'
+import '../../../App.css'
 
 import NavBar from '../Navigation/NavBar';
 import InformationContainer from '../Information/InformationContainer';
